--- front-end/.github/ISSUE_TEMPLATE/❌-bug.md
@@ -1,21 +1,0 @@
----
-name: "❌ Bug"
-about: 버그 사항을 입력해주세요.
-title: "[FE] $페이지 - $발생한 버그"
-labels: "❌ bug"
-assignees: ''
-
----
-
-# 개요
-
-어떠한 버그가 일어났는지 알려주세요.
-
-## 할일
-
-- [ ] TODO1
-- [ ] TODO2
-
-## ETC
-
-기타사항입니다.
