--- front-end/.github/ISSUE_TEMPLATE/🚀-feature.md
@@ -1,21 +1,0 @@
----
-name: "\U0001F680 Feature"
-about: 기능 구현 작업 사항을 입력해주세요.
-title: "[FE] $페이지 - $구현할 기능 이름"
-labels: "\U0001F680 feature"
-assignees: ''
-
----
-
-# 개요
-
-구현할 기능의 내용을 적어주세요.
-
-## 할일
-
-- [ ] TODO1
-- [ ] TODO2
-
-## ETC
-
-기타사항입니다.
