--- front-end/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,12 +1,0 @@
-# 개요
-
-어떠한 이유로 작업을 했는지에 대해서 알려주세요.
-
-## 작업사항
-
-- [ ] TODO1
-- [ ] TODO2
-
-## 기타
-
-기타사항입니다.
