--- front-end/src/assets/README.md
@@ -1,5 +1,0 @@
-# assets
-
-이미지 혹은 폰트와 같은 파일들이 저장되는 폴더입니다.</br>
-이미지와 같은 파일들을 public에 직접 넣는 경우도 있는데 둘의 차이는 컴파일시에 필요한지 여부입니다.</br>
-파비콘과 같이 index.html내부에서 직접 사용하여 컴파일 단계에서 필요하지 않은 파일들은 public에 반면, 컴포넌트 내부에서 사용하는 이미지 파일인 경우 이 assets 폴더에 위치시켜야 합니다.
