+++ front-end/src/components/Main/Card/index.jsx
