+++ front-end/src/components/Main/Card/index.styled.js
