--- front-end/src/components/README.md
@@ -1,4 +1,0 @@
-# components
-
-재사용 가능한 컴포넌트들이 위치하는 폴더입니다.</br>
-컴포넌트는 매우 많아질 수 있기 때문에 이 폴더 내부에서 하위폴더로 추가로 분류하는 경우가 많습니다.
