--- front-end/src/pages/README.md
@@ -1,3 +1,0 @@
-# pages
-
-react router등을 이용하여 라우팅을 적용할 때 페이지 컴포넌트를 이 폴더에 위치시킵니다.</br>
