--- front-end/src/styles/README.md
@@ -1,5 +1,0 @@
-# styles
-
-css 파일들이 포함되는 폴더입니다.</br>
-글로벌 설정, 테마, 폰트, 등을 설정하는 곳입니다.</br>
-만약 개인의 컴포넌트의 스타일을 적용하고 싶으시다면 해당 컴포넌트 폴더에 `CSS` or `CSS in JS` 파일을 생성하여 관리해주세요.
